requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
Werkzeug==3.0.1
edge-tts==6.1.9
yt-dlp==2024.8.6
openai-whisper==20231117
pydub==0.25.1
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
ffmpeg-python==0.2.0